3-autoscaling-2.0.0-18

+++ eat/mouth.go
@@ -69,7 +69,7 @@ func (w *Mouth) Handler(inputInst interface{}) (interface{}, error) {
 		endTime := time.Now()
 		// feedStomach.Pieces = prime_number_list
 		// feedStomach.Size   = input.EndPoint - input.StartPoint + 1
-		feedStomach.Size = len(prime_number_list)
+		feedStomach.Size = int64(len(prime_number_list))
 		feedStomach.EatTime = float64((endTime.Sub(startTime))*time.Millisecond)
 		feedStomach.Cmd = "count"
